Add embedded example sentence support
Adds a pseudo-dictionary called "Embedded" that acts as a source for sentences if the checkbox is enabled.

Need to check whether this interferes with the operation of the software with the checkbox uncheked, and also move it to a better location, allowing the user to specify which column holds the example.